Refactor range processing
